--- a/test/Microsoft.Framework.ApplicationHost.FunctionalTests/ApplicationHostTestCollection.cs
+++ b/test/Microsoft.Framework.ApplicationHost.FunctionalTests/ApplicationHostTestCollection.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using Microsoft.Framework.CommonTestUtils;
+using Xunit;
+
+namespace Microsoft.Framework.ApplicationHost.FunctionalTests
+{
+    [CollectionDefinition("ApplicationHostTestCollection")]
+    public class ApplicationHostTestCollection: ICollectionFixture<DnxRuntimeFixture>
+    {
+    }
+}