… test/Microsoft.Framework.PackageManager.FunctionalTests/DnuWrapTests.cs
… test/Microsoft.Framework.PackageManager.FunctionalTests/PackageManagerFunctionalTestCollection.cs
… test/Microsoft.Framework.PackageManager.FunctionalTests/PackageManagerFunctionalTestFixture.cs