Today, runtime package won't be repeatedly unzip during `build verify`. However it is not the case if the test is run from Visual Studio or run from command line using `dnx . text`. This change leverages the xunit's test collection feature to ensure one fixture is shared in on test assembly among the tests marked in same collection. 

On my machine it reduces the test running time:
```
from:
   Bootstrapper.FunctionalTests  Total: 24, Errors: 0, Failed: 0, Skipped: 0, Time: 99.213s
to:
   Bootstrapper.FunctionalTests  Total: 24, Errors: 0, Failed: 0, Skipped: 0, Time: 58.791s

from:
   Microsoft.Framework.ApplicationHost.FunctionalTests  Total: 20, Errors: 0, Failed: 0, Skipped: 0, Time: 47.367s
to:
   Microsoft.Framework.ApplicationHost.FunctionalTests  Total: 20, Errors: 0, Failed: 0, Skipped: 0, Time: 15.482s

from:
   Microsoft.Framework.PackageManager.FunctionalTests  Total: 82, Errors: 0, Failed: 0, Skipped: 1, Time: 160.156s
to:
   Microsoft.Framework.PackageManager.FunctionalTests  Total: 82, Errors: 0, Failed: 0, Skipped: 1, Time: 121.718s
```

Addressing: https://github.com/aspnet/dnx/issues/1651