@muratg 2015-04-24
Would this help in other repos with long running functional tests?