@troydai 2015-04-24
Unzip runtime is a unique issue for DNX functional test. But if functional test for other repos also have repeating process taking up a lot of time, the `CollectionFixture` can help.