@ChengTian 2015-04-24
:shipit: , thanks for resolving conflicts with my changes :smile: