@dougbu 2015-04-24
@troydai to avoid confusing people reading only commit descriptions, please refer instead to the fixtures.  collections in themselves tend to slow things down (requiring everything in the collection to run serially) but the fixture avoids repetitious setup.