@troydai 2015-04-24
@dougbu thank you for the suggestion, I'll amend the description in the commit.